# go-ecommerce-api/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "A RESTful E-commerce API built with Go and Gin",
        "title": "E-commerce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/admin/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all orders from all users (Admin only)",
                "tags": [
                    "admin"
                ],
                "summary": "Get all orders (Admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.OrdersResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of an order (Admin only)",
                "tags": [
                    "admin"
                ],
                "summary": "Update order status (Admin)",
                "parameters": [
                    {
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.UpdateOrderStatusRequest"
                            }
                        }
                    },
                    "description": "Status update",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all orders for the authenticated user",
                "tags": [
                    "orders"
                ],
                "summary": "Get user's orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.OrdersResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new order from the user's cart",
                "tags": [
                    "orders"
                ],
                "summary": "Create order from cart",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.CreateOrderRequest"
                            }
                        }
                    },
                    "description": "Order creation request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.OrderResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create order and payment intent in one step. Cart is automatically cleared after successful checkout.",
                "tags": [
                    "orders"
                ],
                "summary": "Checkout and create payment",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.CheckoutResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific order by its ID",
                "tags": [
                    "orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.OrderResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "models.Cart": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.CartItem"
                        }
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/models.User"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "models.CartItem": {
                "type": "object",
                "properties": {
                    "cart": {
                        "$ref": "#/components/schemas/models.Cart"
                    },
                    "cart_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "product": {
                        "$ref": "#/components/schemas/models.Product"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
...
            "models.CheckoutOrderResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "shipping": {
                        "type": "number",
                        "example": 5.99
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "total": {
                        "type": "number",
                        "example": 292.64
                    }
                }
            },
            "models.CheckoutPaymentResponse": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number",
                        "example": 292.64
                    },
                    "currency": {
                        "type": "string",
                        "example": "usd"
                    },
                    "payment_intent_id": {
                        "type": "string",
                        "example": "pi_3SZA4XIH7jU9U4pR0VhsSo69"
                    },
                    "status": {
                        "type": "string",
                        "example": "requires_payment_method"
                    }
                }
            },
            "models.CheckoutResponse": {
                "type": "object",
                "properties": {
                    "client_secret": {
                        "type": "string",
                        "example": "pi_xxx_secret_yyy"
                    },
                    "message": {
                        "type": "string",
                        "example": "Checkout successful"
                    },
                    "order": {
                        "$ref": "#/components/schemas/models.CheckoutOrderResponse"
                    },
                    "payment": {
                        "$ref": "#/components/schemas/models.CheckoutPaymentResponse"
                    }
                }
            },
            "models.CreateOrderRequest": {
                "type": "object",
                "properties": {
                    "shipping": {
                        "type": "number",
                        "minimum": 0,
                        "example": 10
                    }
                }
            },
            "models.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "models.Order": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.OrderItem"
                        }
                    },
                    "payment_intent_id": {
                        "type": "string"
                    },
                    "payment_method": {
                        "description": "card, cash, etc.",
                        "type": "string"
                    },
                    "payment_status": {
                        "description": "pending, processing, succeeded, failed, cancelled",
                        "type": "string"
                    },
                    "shipping": {
                        "type": "number"
                    },
                    "status": {
                        "description": "pending, processing, shipped, delivered, cancelled",
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    },
                    "transaction_id": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/models.User"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "models.OrderItem": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "order": {
                        "$ref": "#/components/schemas/models.Order"
                    },
                    "order_id": {
                        "type": "integer"
                    },
                    "price": {
                        "description": "Price at time of order",
                        "type": "number"
                    },
                    "product": {
                        "$ref": "#/components/schemas/models.Product"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "models.OrderResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/models.Order"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "models.OrdersResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Order"
                        }
                    }
                }
            },
...
            "models.Product": {
                "type": "object",
                "properties": {
                    "badge": {
                        "type": "string"
                    },
                    "badge_color": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "featured": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "original_price": {
                        "type": "number"
                    },
                    "price": {
                        "type": "number"
                    },
                    "rating": {
                        "type": "number"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
...
            "models.SuccessResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "models.UpdateOrderStatusRequest": {
                "type": "object",
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "shipped"
                    }
                }
            },
            "models.User": {
                "type": "object",
                "properties": {
                    "cart": {
                        "$ref": "#/components/schemas/models.Cart"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Order"
                        }
                    },
                    "role": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "wishlist": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Wishlist"
                        }
                    }
                }
            },
...
            "models.Wishlist": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "product": {
                        "$ref": "#/components/schemas/models.Product"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/models.User"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
...
        }
    }
}